+++ server/schema/UserResolvers.ts
@@ -1,61 +1,80 @@
 /* eslint-disable no-underscore-dangle */
 import { IResolvers } from '@graphql-tools/utils';
 import bcrypt from 'bcrypt';
+import { ApolloError } from 'apollo-errors';
 
 import UserDB from '../models/User';
 import ProjectDB from '../models/Project';
 import QueryDB from '../models/Query';
 import sessions from '../utilities/sessions';
+import { WrongCredentialsError } from './errors';
+import { MongoProject, MongoUser } from '../../@types/resolver';
 
 const resolvers: IResolvers = {
     /*
      * To query nested Project Data in User Object
      */
     User: {
-        projects: (parent: User): Promise<Project[] | Error> =>
-            ProjectDB.find({ userID: parent.id })
-                .then((projects: Project[]): Project[] => projects)
-                .catch((err: Error): Error => new Error(`DB query failed: ${err}`)),
+        projects: (parent: User, args, context: Context): Promise<Project[] | Error> | Error => {
+            const { authenticated } = context;
+            if (authenticated === false) return new Error('Unauthorized to make this request');
+            return ProjectDB.find({ userID: parent.id })
+                .then((projects: MongoProject[]): Project[] => projects)
+                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
+        },
 
-        // FIXME: What is the use case for the project field on user?
-        // project: (parent: User, args: QueryByID): Promise<Project | Error> =>
-        //     ProjectDB.findOne({ _id: args.id, userID: parent.id })
-        //         .then((project: Project): Project => {
+        // FIXME: What's the use case for this
+        // project: (
+        //     parent: User,
+        //     args: QueryByID,
+        //     context: Context
+        // ): Promise<Project | Error> | Error => {
+        //     const { authenticated } = context;
+        //     if (authenticated === false) return new Error('Unauthorized to make this request');
+        //     return ProjectDB.findOne({ _id: args.id, userID: parent.id })
+        //         .then((project: any): Project => {
         //             if (!project) throw new Error('Project not found');
         //             return project;
         //         })
-        //         .catch((err: Error): Error => new Error(`DB query failed: ${err}`)),
+        //         .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
+        // },
     },
 
     Query: {
         // FIXME: What is the business case for returning a list of users?
-        users: (): Promise<User[] | Error> =>
-            UserDB.find()
-                .then((users: User[]): User[] => users)
-                .catch((err: Error): Error => new Error(`DB query failed: ${err}`)),
-
-        user: (parent: undefined, args: QueryByID): Promise<User | Error> => {
-            const { id } = args;
-
-            return UserDB.findOne({ _id: id })
-                .then((user: User): User => {
-                    if (!user) throw new Error('User does not exist');
-                    return user;
-                })
-                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
-        },
-
-        // Auth Queries
         checkAuth: (parent, args, context): Promise<User | Error> | null => {
             if (context.authenticated) {
                 return UserDB.findOne({ _id: context.user.id })
-                    .then((user: User): User => {
+                    .then((user: MongoUser): User => {
                         if (!user) throw new Error('User does not exist');
                         return user;
                     })
                     .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
             }
             return null;
+        },
+        users: (parent, args, context: Context): Promise<User[] | Error> | Error => {
+            const { authenticated } = context;
+            if (authenticated === false) return new Error('Unauthorized to make this request');
+            return UserDB.find()
+                .then((users: MongoUser[]): User[] => users)
+                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
+        },
+        user: (
+            parent: undefined,
+            args: QueryByID,
+            context: Context
+        ): Promise<User | Error> | Error => {
+            const { authenticated } = context;
+            if (authenticated === false) return new Error('Unauthorized to make this request');
+            const { id } = args;
+
+            return UserDB.findOne({ _id: id })
+                .then((user: MongoUser): User => {
+                    if (!user) throw new Error('User does not exist');
+                    return user;
+                })
+                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
         },
     },
 
@@ -65,33 +84,31 @@ const resolvers: IResolvers = {
          */
         login: async (parent: undefined, args: GetUserArgs): Promise<User | Error> => {
             const { email, password } = args.user;
-
             return UserDB.findOne({ email })
-                .then(async (user: any): Promise<User> => {
-                    if (!user.email) {
-                        throw new Error('Email not found.');
-                    }
+                .then(async (user: MongoUser): Promise<User> => {
+                    if (!user) throw new WrongCredentialsError();
                     const verifyPassword: boolean = await bcrypt.compare(password, user.password);
                     if (!verifyPassword) {
-                        throw new Error('Password you entered is incorrect.');
+                        throw new WrongCredentialsError();
                     }
-
-                    const token = sessions.create({ id: user._id });
+                    const token = sessions.create({ id: user.id });
                     return {
                         token,
                         email: user.email,
                         password: user.password,
-                        id: user._id,
-                        projects: [],
+                        id: user.id,
                     };
                 })
-                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
+                .catch((err: Error): Error => {
+                    if (err instanceof ApolloError) throw new Error(`${err}`);
+                    else throw new Error('Try Again Later');
+                });
         },
         signup: async (parent: undefined, args: CreateUserArgs): Promise<User | Error> => {
             const { email, password } = args.user;
             const hash: string = await bcrypt.hash(password, 11);
             return UserDB.findOne({ email })
-                .then(async (user: User): Promise<User> => {
+                .then(async (user: MongoUser): Promise<User> => {
                     if (user) throw new Error('User already exists');
                     const newUser = new UserDB({
                         email,
@@ -99,21 +116,27 @@ const resolvers: IResolvers = {
                         projects: [],
                     });
                     const savedUser = await newUser.save();
-                    if (!savedUser) throw new Error('Could not save user. Try again later.');
-                    const token = sessions.create({ id: savedUser._id });
+                    if (!savedUser) throw new Error('Try again later.');
+                    const token = sessions.create({ id: savedUser._id.toString() });
+                    const idString = savedUser._id.toString();
+
                     return {
                         token,
                         email: savedUser.email,
                         password: savedUser.password,
-                        id: savedUser._id,
-                        projects: [],
+                        id: idString,
                     };
                 })
-                .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
+                .catch((err: Error): Error => new Error(`${err}`));
         },
-        updateUser: async (parent: undefined, args: UpdateUserArgs): Promise<User | Error> => {
+        updateUser: async (
+            parent: undefined,
+            args: UpdateUserArgs,
+            context: Context
+        ): Promise<User | Error> => {
             const { id, email, password } = args.user;
-
+            const { authenticated } = context;
+            if (authenticated === false) return new Error('Unauthorized to make this request');
             // update object depends on what user passes into the mutation variables
             const updateBody: object = {};
 
@@ -124,28 +147,34 @@ const resolvers: IResolvers = {
             }
 
             return UserDB.findByIdAndUpdate(id, updateBody, { new: true })
-                .then((user: User): User => {
+                .then((user: MongoUser): User => {
                     if (!user) throw new Error('User not found');
                     return user;
                 })
                 .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
         },
-        deleteUser: async (parent: undefined, args: QueryByID): Promise<User | Error> => {
+        deleteUser: async (
+            parent: undefined,
+            args: QueryByID,
+            context: Context
+        ): Promise<User | Error> => {
             const { id } = args;
+            const { authenticated } = context;
+            if (authenticated === false) return new Error('Unauthorized to make this request');
 
             // delete each user project query from DB
             await QueryDB.deleteMany({ userID: id }).catch(
-                (err) => `DB user project query deletion failed ${err}`
+                (err: Error) => `DB user project query deletion failed ${err}`
             );
 
             // delete each user project from DB
             await ProjectDB.deleteMany({ userID: id }).catch(
-                (err) => `DB user project deletion failed ${err}`
+                (err: Error) => `DB user project deletion failed ${err}`
             );
 
             // deletes user from DB and returns user object
             return UserDB.findByIdAndRemove(id)
-                .then((user: User): User => user)
+                .then((user: MongoUser): User => user)
                 .catch((err: Error): Error => new Error(`DB query failed: ${err}`));
         },
     },
